+++ support/cas-server-support-passwordless-mongo/src/main/java/org/apereo/cas/impl/token/MongoDbPasswordlessTokenRepository.java
@@ -40,7 +40,7 @@ public class MongoDbPasswordlessTokenRepository extends BasePasswordlessTokenRep
     private final PasswordlessAuthenticationMongoDbTokensProperties properties;
 
     public MongoDbPasswordlessTokenRepository(final PasswordlessAuthenticationMongoDbTokensProperties properties,
-                                              final int tokenExpirationInSeconds,
+                                              final long tokenExpirationInSeconds,
                                               final CipherExecutor cipherExecutor,
                                               final MongoOperations mongoTemplate) {
         super(tokenExpirationInSeconds, cipherExecutor);
